Authentication middleware developed and implemented
