+++ src/controllers/user.controller.js
@@ -15,28 +15,6 @@ export const generateToken = (req, res) => {
   }
 };
 
-export const login = async (req, res, next) => {
-  const { email, password } = req.body;
-  try {
-    const user = await prisma.user.findFirst({
-      where: {
-        email: email,
-      },
-    });
-    const isValidUser = bcrypt.compareSync(password, user.password);
-    if (isValidUser) {
-      //aqui va el next para la generacion del token
-      next();
-    } else {
-      res
-        .status(401)
-        .json({ error: true, message: "invalid user or password" });
-    }
-  } catch (error) {
-    res.status(500).json({ error: true });
-  }
-};
-
 export const register = async (req, res) => {
   const { email, password } = req.body;
   const hash = bcrypt.hashSync(password, 12);
