+++ src/routes/user.routes.js
@@ -1,9 +1,6 @@
 import express from "express";
-import {
-  login,
-  register,
-  generateToken,
-} from "../controllers/user.controller.js";
+import { register, generateToken } from "../controllers/user.controller.js";
+import { login } from "../middlewares/login.middle.js";
 
 const router = express.Router();
